chore: move dashboard pages to their own folder
