+++ lifesheet-fe/src/App.tsx
@@ -4,17 +4,17 @@ import { IntlProvider, FormattedMessage } from 'react-intl';
 import { useEffect, useState } from 'react';
 
 import './App.css';
-import { CVData } from './components/cv-data';
+import { CVData } from './components/dashboard/cv-data';
 import { Welcome } from './components/welcome';
 import cvsService from './services/cvs-service';
 import userService from './services/user-service';
-import { TailorCV } from './components/tailor-cv';
-import { PlansPage } from './components/plans/plans-page';
-import { CheckoutPage } from './components/plans/checkout-page';
-import { CheckoutSuccessPage } from './components/plans/checkout-success';
-import { CheckoutCancelPage } from './components/plans/checkout-cancel';
-import { ExportPdf } from './components/export-pdf';
-import TailoredCVs from './components/tailored-cvs';
+import { TailorCV } from './components/dashboard/tailor-cv';
+import { PlansPage } from './components/dashboard/plans/plans-page';
+import { CheckoutPage } from './components/dashboard/plans/checkout-page';
+import { CheckoutSuccessPage } from './components/dashboard/plans/checkout-success';
+import { CheckoutCancelPage } from './components/dashboard/plans/checkout-cancel';
+import { ExportPdf } from './components/dashboard/export-pdf';
+import TailoredCVs from './components/dashboard/tailored-cvs';
 import { Onboarding } from './components/onboarding';
 import { Dashboard } from './components/dashboard';
 import { LoadingIndicator } from './components/ui/loading-indicator';
